requirements.txt
transformers==4.32.0
pyyaml==5.3.1
gdown==4.7.1
torch==1.11.0
telebot
zipfile36
